requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.0.0
python-dotenv>=1.0.0
requests>=2.28.0
requests-oauthlib>=1.3.0
sqlalchemy>=2.0.0
pymysql>=1.1.0

